Add files via upload
This is the current working api/database as of 4/8/18. users, listings, and books seem to be linked, but more testing is necessary. From here, searching books and filtering listings must be implemented, and a couple other specific functions for the front-ends sake should be added (for example, if a user is searched for by google_token, all of that user's listings, and all book information for each listing, should be sent back as one JSON object, since the only time a user is ever searched for by google_token is when a user logs into their home page (there is no function for searching for your friend's postings, at least for mvp)).